requirements.txt
newspaper
pika
python-dateutil
python-jsonrpc
pymongo
redis
requests
sklearn
watchdog
